+++ seminarimport.php
@@ -12,10 +12,32 @@ if(!empty($_FILES['file']['name']))
  //$query_5='truncate table customer_table';
  //$statement = $connect->prepare($query_5);
  //$statement->execute();
+ 
+ $arr=array();
+ $count=0;
+ $sql3="show tables";
+ $result = mysqli_query($con,$sql3);
+ while($table = mysqli_fetch_array($result))
+ {
+ 	$arr[$count]=$table[0];
+ 	$count=$count+1;
+ }
 
- $sql2='drop table '.$filename_without_ext.'';
- $statement1 = $connect->prepare($sql2);
- $statement1->execute();
+ $count1=0;
+	 while($count1<=$count)
+	 {
+	 	if ($arr[$count1]==$filename_without_ext) 
+	 	{
+	 		$sql2='drop table '.$filename_without_ext.'';
+	 		$statement1 = $connect->prepare($sql2);
+	 		$statement1->execute();
+	 		break;
+	 	}
+	 	else
+	 	{
+	 		$count1=$count1+1;
+	 	}
+	 }
 
  $sql1='create table '.$filename_without_ext.' as select * from test4';
  $statement1 = $connect->prepare($sql1);
